fix: resolve off-by-one error and cleanup types
- Fix off-by-one error as reported in #184.
- Refactor: Remove duplicate const declaration to fix shadowing and
  remove unnecessary type casting in openxr\src\hand_tracker.rs.
- Chore: Add *.dll to .gitignore.

Closes #184

+++ openxr/src/hand_tracker.rs
@@ -8,7 +8,7 @@ pub use sys::HandJointEXT as HandJoint;
 pub use sys::HandJointLocationEXT as HandJointLocation;
 pub use sys::HandJointVelocityEXT as HandJointVelocity;
 
-pub const HAND_JOINT_COUNT: usize = sys::HAND_JOINT_COUNT_EXT as usize;
+pub const HAND_JOINT_COUNT: usize = sys::HAND_JOINT_COUNT_EXT;
 
 pub struct HandTracker {
     pub(crate) session: Arc<session::SessionInner>,
